# api/openapi.yaml
openapi: 3.1.0
info:
  contact:
    name: CDE Platform engineering
  description: |
    Common Data Environment API for construction and asset information.

    ## Tenant scoping

    Every request is scoped to exactly one tenant, and the tenant is taken **from the authenticated principal** — never from a path, query, header, or body parameter. There is no tenant selector on any endpoint, and supplying one has no effect. A resource belonging to another tenant is not visible: reading it by id returns `404`, not `403`, so the API does not confirm that an id exists somewhere else.

    ## Errors

    Every error response is an RFC 9457 Problem Details document with the media type `application/problem+json`. It carries a `traceId` that identifies the request in the server logs — quote it when reporting a problem. The `type` member is a stable relative URI identifying the kind of failure, and is safe to branch on when the status code alone is too coarse; `detail` is human-readable text whose wording may change, and is not.

    Validation failures return `422` with an `invalidFields` array naming each rejected field. A request body that cannot be parsed at all returns `400`, because no field can be named.

    ## Rate limits

...

    ## Tracing

    Every response carries `X-Trace-Id`. Send a W3C `traceparent` header and its trace id is adopted, so a client-side trace and the server-side logs share one identifier.
  license:
    name: Proprietary
  title: CDE Platform API
...
servers:
...
security:
- bearerAuth: []
...
paths:
...
  /api/viewer/{documentId}:
    get:
      description: |-
        Returns whatever suits the document's format: SVG markup for a drawing, a pointer to the bytes for a PDF, extracted geometry for a model, or an explanation of why it cannot be opened. Read the `type` member first — it decides the shape of the rest.

        Images and mesh formats (GLB, STL, OBJ, PLY, DAE) are returned as raw bytes with the matching content type rather than as JSON, so check the response's content type before parsing it. Office documents are converted and returned as PDF bytes.

        **A `200` does not mean the document opened.** The variants that report a problem — `error`, `dwg_binary`, `office_error`, `revit_binary`, `3d_error`, `unsupported` — are also returned with `200`. This is a defect kept for compatibility: every client branches on `type`, so correcting the status is a coordinated change on both sides.

        Requires the `document:read` permission.
      operationId: getViewerPayload
      parameters:
      - description: Identifier of the document to open.
        example: 1180
        in: path
        name: documentId
        required: true
        schema:
          type:
          - integer
          format: int64
          example: null
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ViewerPayload"
                example: null
            application/pdf:
              schema:
                format: binary
                example: null
            image/*:
              schema:
                format: binary
                example: null
            model/gltf-binary:
              schema:
                format: binary
                example: null
          description: "What the viewer needs, or a described reason it cannot open\
            \ the document."
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: "The request could not be read — malformed JSON, or a value\
            \ outside an accepted set. No field can be named, because parsing did\
            \ not get that far."
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: "No credential was presented, or the one presented is expired\
            \ or invalid."
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: "Authenticated, but without the permission this operation requires."
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: No document with that id is visible to the caller.
        "429":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: "A rate limit was exceeded. `Retry-After` gives the seconds\
            \ to wait. Limits apply per account, per source address and per tenant\
            \ independently."
        "500":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: The request failed for a reason the caller cannot act on. Quote
            the `traceId` when reporting it.
      summary: Get what the viewer needs to open a document
      tags:
      - Viewer
...
  /api/viewer3d/{documentId}:
    get:
      description: |-
        IFC is extracted by the conversion service and returned as JSON. Mesh formats — GLB, glTF, OBJ, STL, PLY, DAE — are streamed as raw bytes with the matching content type and an `X-3D-Format` header, so check the content type before parsing. Revit files cannot be opened at all and report why.

        As with the document viewer, a problem is reported with `200` and a described payload rather than an error status. Read `success` and `type`.

        The structured hierarchy from `/tree` is the primary route to a model's information; this endpoint serves the visual layer over it.

        Requires the `document:read` permission.
      operationId: getModel
      parameters:
      - description: Identifier of the model document.
        example: 1212
        in: path
        name: documentId
        required: true
        schema:
          type:
          - integer
          format: int64
          example: null
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ViewerPayload"
                example: null
            model/gltf-binary:
              schema:
                format: binary
                example: null
          description: "The geometry, or a described reason it is unavailable."
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: "The request could not be read — malformed JSON, or a value\
            \ outside an accepted set. No field can be named, because parsing did\
            \ not get that far."
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: "No credential was presented, or the one presented is expired\
            \ or invalid."
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: "Authenticated, but without the permission this operation requires."
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: No document with that id is visible to the caller.
        "429":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: "A rate limit was exceeded. `Retry-After` gives the seconds\
            \ to wait. Limits apply per account, per source address and per tenant\
            \ independently."
        "500":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: The request failed for a reason the caller cannot act on. Quote
            the `traceId` when reporting it.
      summary: Get a model's geometry
      tags:
      - 3D and model viewer
  /api/viewer3d/{documentId}/tree:
    get:
      description: |-
        The structured route to everything the model holds — the interface a keyboard or a screen reader uses, and the one a rendered canvas cannot replace. It supports the same navigation and selection the visual view does.

        **When the conversion service cannot be reached, a placeholder outline is returned instead of the model's real hierarchy**, and its root carries `synthetic: true`. It lists the element classes a model of this kind usually contains, not this model's contents. A client must not present it as the model's structure; check the flag.

        Requires the `document:read` permission.
      operationId: getModelTree
      parameters:
      - description: Identifier of the model document.
        example: 1212
        in: path
        name: documentId
        required: true
        schema:
          type:
          - integer
          format: int64
          example: null
      responses:
        "200":
          content:
            application/json:
              schema:
                example: null
                items:
                  $ref: "#/components/schemas/ModelTreeNode"
                  example: null
          description: "The hierarchy, rooted at the model. Check the root's `synthetic`\
            \ flag before treating it as the model's real structure."
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: "The request could not be read — malformed JSON, or a value\
            \ outside an accepted set. No field can be named, because parsing did\
            \ not get that far."
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: "No credential was presented, or the one presented is expired\
            \ or invalid."
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: "Authenticated, but without the permission this operation requires."
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: No document with that id is visible to the caller.
        "429":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: "A rate limit was exceeded. `Retry-After` gives the seconds\
            \ to wait. Limits apply per account, per source address and per tenant\
            \ independently."
        "500":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: The request failed for a reason the caller cannot act on. Quote
            the `traceId` when reporting it.
      summary: Get a model's hierarchy as a navigable tree
      tags:
      - 3D and model viewer
components:
  schemas:
...
    ModelTreeNode:
      description: "One node of a model's hierarchy. The tree is the accessible route\
        \ to the model's information: it must support the same navigation, search\
        \ and selection the rendered view does, because a canvas on its own cannot."
      example: null
      properties:
        children:
          type:
          - array
          description: Elements contained by this one.
          example: null
          items:
            $ref: "#/components/schemas/ModelTreeNode"
            example: null
        expanded:
          type:
          - boolean
          description: Whether a client should show this node's children initially.
          example: false
        id:
          type:
          - string
          description: Identifier of this element within the model.
          example: 3Vw$Kq2Kn0zeF1yZ8bQwXt
        name:
          type:
          - string
          description: Name to show for this element.
          example: Basic Wall — 215mm
        selected:
          type:
          - boolean
          description: Whether this node is currently selected.
          example: false
        synthetic:
          type:
          - boolean
          description: |-
            Set only on the root, and only when the hierarchy could not be extracted from the model itself — the conversion service was unreachable, so what is returned is a placeholder outline of the element classes a model of this kind usually contains, not this model's contents.

            Absent or false means the hierarchy came from the model. A client must not present a synthetic tree as the model's structure.
          example: true
        type:
          type:
          - string
          description: The element's class in the source model.
          example: IfcWall
        visible:
          type:
          - boolean
          description: Whether this element is currently shown in the rendered view.
          example: true
      required:
      - id
      - name
      - type
...
    PdfPayload:
      description: A PDF. The bytes are fetched separately so the viewer can stream
        them; this carries the pointer and the metadata around it.
      example: null
      properties:
        drawingNumber:
          type:
          - string
          description: "Drawing number from the title block, empty when none is held."
          example: RVD-XX-02-DR-A-1200
        fileName:
          type:
          - string
          description: Name of the underlying file.
          example: RVD-XX-02-DR-A-1200.pdf
        name:
          type:
          - string
          description: The document's display name.
          example: GA Plan — Level 02
        pdfUrl:
          type:
          - string
          format: uri-reference
          description: "Where to fetch the bytes. The version is in the query string\
            \ because processing replaces the bytes behind a fixed URL, and without\
            \ it a browser would serve the previous version from cache."
          example: /api/viewer/1180/pdf?v=3
        revision:
          type:
          - string
          description: "Revision identifier, empty when none is held."
          example: P02.1
        type:
          type:
          - string
          description: Always `pdf`.
          enum:
          - pdf
          example: pdf
        version:
          type:
          - integer
          format: int32
          description: Version the pointer refers to.
          example: 3
          minimum: 1
      required:
      - name
      - pdfUrl
      - type
      - version
    ProblemDetail:
      type:
      - object
      description: "RFC 9457 Problem Details. The single error shape returned by every\
        \ endpoint, for every status."
      example: null
      properties:
        detail:
          type:
          - string
          description: "What happened and what to do next, written for a person to\
            \ read. Never contains a stack trace, SQL, or an internal type name."
          example: 1 field is invalid.
        instance:
          type:
          - string
          format: uri-reference
          description: Path of the request that failed.
          example: /api/projects
        invalidFields:
          type:
          - array
          description: "Present on `422` validation failures. Names every field that\
            \ was rejected, so a client can attach each message to the control it\
            \ belongs to rather than to the form as a whole."
          example: null
          items:
            type:
            - object
            description: One field the request was rejected for.
            example: null
            properties:
              field:
                type:
                - string
                description: "Path of the rejected field, as it appears in the request\
                  \ body."
                example: name
              message:
                type:
                - string
                description: "What is wrong with the value, phrased for the person\
                  \ who typed it."
                example: must not be blank
            required:
            - field
            - message
        status:
          type:
          - integer
          format: int32
          description: "HTTP status code, repeated here so a logged body is self-contained."
          example: 422
          maximum: 599
          minimum: 400
        timestamp:
          type:
          - string
          format: date-time
          description: "When the failure was recorded, UTC."
          example: 2026-02-17T09:41:12.884Z
        title:
          type:
          - string
          description: Short summary of the problem type. Does not vary between occurrences.
          example: Validation failed
        traceId:
          type:
          - string
          description: "Identifies this request in the server logs, and is also returned\
            \ in the `X-Trace-Id` header. Quote it when reporting a problem."
          example: 4f8a1c2e9b7d6a5f3e2d1c0b9a8f7e6d
          pattern: "^[0-9a-f]{32}$"
        type:
          type:
          - string
          format: uri-reference
          description: "Stable identifier for this kind of problem, relative to the\
            \ API host. Safe to branch on — unlike `detail`, its value does not change\
            \ with wording."
          example: /problems/validation-failed
      required:
      - type
      - title
      - status
      - detail
      - traceId
...
    SvgPayload:
      description: "A drawing, as SVG markup ready to render. Used for uploaded SVG\
        \ and for CAD converted to it."
      example: null
      properties:
        content:
          type:
          - string
          description: The SVG markup itself.
          example: <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 841 595"/>
        drawingNumber:
          type:
          - string
          description: "Drawing number from the title block, empty when none is held."
          example: RVD-XX-02-DR-A-1200
        name:
          type:
          - string
          description: The document's display name.
          example: GA Plan — Level 02
        renderedBy:
          type:
          - string
          description: "Which converter produced the markup. Present for converted\
            \ CAD only, and useful when a drawing renders oddly — the fallback renderer\
            \ supports less than the primary one."
          example: ezdxf
        revision:
          type:
          - string
          description: "Revision identifier, empty when none is held."
          example: P02.1
        type:
          type:
          - string
          description: Always `svg`.
          enum:
          - svg
          example: svg
      required:
      - content
      - name
      - type
...
    ViewerErrorPayload:
      description: "The document could not be read at all — no file recorded, or the\
        \ file is missing from storage. Returned with `200`; read `type`."
      example: null
      properties:
        error:
          type:
          - string
          description: "What went wrong, phrased for a person."
          example: This document has no file behind it.
        type:
          type:
          - string
          description: Always `error`.
          enum:
          - error
          example: error
      required:
      - error
      - type
    ViewerPayload:
      description: |-
        What the viewer needs to open a document. Read `type` first — it decides which other members are present.

        Note that the failure variants (`error`, `dwg_binary`, `office_error`, `3d_error`, `revit_binary`, `unsupported`) are returned with `200`, so a successful status does not mean the document opened. Branch on `type`.
      discriminator:
        propertyName: type
      example: null
      oneOf:
      - $ref: "#/components/schemas/SvgPayload"
        example: null
      - $ref: "#/components/schemas/PdfPayload"
        example: null
      - $ref: "#/components/schemas/ModelPayload"
        example: null
      - $ref: "#/components/schemas/UnopenablePayload"
        example: null
      - $ref: "#/components/schemas/ViewerErrorPayload"
        example: null
...
  securitySchemes:
    apiKey:
      description: "Machine credential, sent as `X-API-Key`. Shown once at creation\
        \ and stored only as a SHA-256 digest, so a lost key is replaced rather than\
        \ recovered. Scoped, expiring, and rate-limited per key."
      in: header
      name: X-API-Key
      type: apiKey
    bearerAuth:
      bearerFormat: JWT
      description: "Signed JSON Web Token from `POST /api/auth/login`, sent as `Authorization:\
        \ Bearer <token>`. The token carries the subject and the tenant; the tenant\
        \ claim is authoritative and is not overridable by any request parameter."
      scheme: bearer
      type: http
    oauth2:
      description: Authorization code with PKCE for user-facing clients; client credentials
        for machine clients. The implicit flow is not offered.
      flows:
        authorizationCode:
          authorizationUrl: /oauth2/authorize
          refreshUrl: /oauth2/token
          scopes:
            annotation:read: Read markup on a document
            annotation:write: "Create, update and resolve markup"
            document:process: Run operations that rewrite a document
            document:read: Read document metadata and content
            document:write: "Upload, replace and delete documents"
            project:read: List and read projects
            project:write: "Create, update and delete projects"
            signature:read: Read and verify digital signatures
            signature:write: Apply a digital signature to a document version
          tokenUrl: /oauth2/token
        clientCredentials:
          scopes:
            annotation:read: Read markup on a document
            annotation:write: "Create, update and resolve markup"
            document:process: Run operations that rewrite a document
            document:read: Read document metadata and content
            document:write: "Upload, replace and delete documents"
            project:read: List and read projects
...